Challenge unlock explosions (from character select screen)
* Generalised the existing system to work here, no need to copypaste
* Spawned on unlock AND on debug button, so you can watch how they propogate
* Currently unsure on how large tile does it - right now it just spawns two sets of explosions on topleft and bottomright corners.

Also fixed an issue where a large tile could fail to show its highlight border if it looped between the two ends of the challenge grid

+++ src/k_menu.h
@@ -653,9 +653,9 @@ extern UINT8 setup_maxpage;
 #define CSEXPLOSIONS 48
 
 extern struct setup_explosions_s {
-	UINT8 x, y;
+	UINT16 x, y;
 	UINT8 tics;
-	UINT8 color;
+	UINT16 color;
 } setup_explosions[CSEXPLOSIONS];
 
 typedef enum
